backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CrewAI and AI dependencies
crewai==0.28.8
groq==0.4.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

requests==2.31.0

# Additional utilities
typing-extensions==4.8.0
